audio: moved InMemoryResource in own file
