+++ cpp/audio/resource.cpp
@@ -2,7 +2,7 @@
 
 #include "resource.h"
 
-#include "in_memory_loader.h"
+#include "in_memory_resource.h"
 #include "../engine.h"
 #include "../util/error.h"
 #include "../log.h"
@@ -41,32 +41,6 @@ std::shared_ptr<Resource> Resource::create_resource(category_t category,
 		return std::make_shared<DynamicResource>(category, id, path, format);
 	default:
 		throw util::Error{"Unsupported loader policy"};
-	}
-}
-
-// in memory resource
-
-InMemoryResource::InMemoryResource(category_t category, int id,
-		const std::string &path, format_t format)
-		:
-		Resource{category, id}  {
-	auto loader = InMemoryLoader::create(path, format);
-	buffer = loader->get_resource();
-}
-
-std::tuple<const int16_t*,uint32_t> InMemoryResource::get_data(
-		uint32_t position, uint32_t data_length) {
-	// if the resource's end has been reached
-	uint32_t length = static_cast<uint32_t>(buffer.size());
-	if (position >= length) {
-		return std::make_tuple(nullptr, 0);
-	}
-
-	const int16_t *buf_pos = &buffer[position];
-	if (data_length > length - position) {
-		return std::make_tuple(buf_pos, length - position);
-	} else {
-		return std::make_tuple(buf_pos, data_length);
 	}
 }
